Reworked mkfs.myfs and added (error) messages.

+++ includes/fat.h
@@ -15,9 +15,9 @@ private:
 
 public:
 	FAT();
-	uint16_t iterateFAT(uint16_t firstBlock, std::list<uint16_t>* list);
+	int iterateFAT(uint16_t firstBlock, std::list<uint16_t>* list);
 	int deleteFromFAT(uint16_t firstBlock);
-	uint16_t addToFAT(uint16_t firstBlock, uint16_t nextAddress);
+	int addToFAT(uint16_t firstBlock, uint16_t nextAddress);
 	void addLastToFAT(uint16_t lastAddress);
 	void setAll(char* p);
 	void getAll(char* p);
